Prototype Final Hardware File
This needs to have an output pin created and assigned in parallel to the AXI register.  Pretty much a modified Lab 4 with changes to the blink_tunable file that allows for LED state to be solid on/off or blinking at 10Hz.  To make this work, need to assign an output pin for the LED that connects to the output signal of blink_tunable.  This was used in lab 4 as the register that gets read from for the digital LED, and can still be read this way but can also have a parallel connection to the camera LED.  **NEEDS TO BE MODIFIED AND COMPILED TO WORK**